Squashed 'lispBM/lispBM/' changes from f18a5c10..85675e29
85675e29 refactor of COND to be able to remove the single-use cddr function.
9d0bb92c added a reminder note in heap.c
e0f7403a small refactor in env and addition of high-level array eq comparison
463e9c32 Change to how reader flags are saved and restored.
8271b1ed fix bug: some constant values were not flattenable.
7bf86cf9 small refactor and added caching of clean-cl-env symbol in case of CLEAN_UP_CLOSURES.
ea3ec7da Marking an array is now an O(N) operation with an additional storage of one 32bit value per array.
872355a4 added a self-reminder in heap.c about how to make array marking linear in size of array.
a0c7afb9 be more explicit in what the INCREMENTA_READ context flag is set to upon launching a read
d71c218a added an example of how to implement map over an array
449075bb added note about garbage collection of arrays (so that I remember)
89bff1c9 updated array mark algorithm and added a note about why it is a bit of a tricky problem to mark arrays
fa1e1e41 adding a note about why GC for high-level arrays is a bit of a tricky problem
384c45d0 array-to-list and list-to-array, length now also returns the size of an array
1f226ec7 fixing two issues in print indicated by sa
9abbeddc clean up in print.c
059295c8 high level arrays can now be stored in flash.
32d162b4 high level arrays can be flattened and unflattened
2f96314c experimental high-level arrays
835e1f22 internal renaming of ARRAY to BYTEARRAY has begun
9e733c42 more cases where symbol decoding has been removed
d23995cf removing a number of encodes/decodes of symbols
c6171530 Version 0.24.0
0d4fb1d0 added some info about index into rest-args
9fb71256 made a few cases of floating point literals in lbmref more readable
9bd5e055 fix typo
0961f4bc update lbmref with more text about rest-args
24959852 a start of a flat-value specification in the lbmref.md
37c9fe3a updated 64 bit tests to follow same approach as run_tests.sh and run_tests_gc.sh. Also added 64bit tests to the release script
0d323069 sdlrepl has been absorbed into the standard repl and can be made by 'make sdl'
37f28977 closing some warnings when doing 64bit build. Some things should be specified better.. such as  allowed range of array sizes
376f10f8 small tweak, cleaning
05326ef2 removed old notes
a89aa19b remove old scripts related to building of previously removed examples
238d4752 removing a bunch of old examples that I will not have the endurance to keep updating
fb741051 examples have been moved into repl.
6d52ae40 tweak typo
c63863a8 rename to md
ad7dacee update README
ef0a75d9 update Makefile for change in directory hierarchy
012a82ea cleaning and updating examples in repl directory
35074ebc improved rendering of examples containing float literals
d558c580 started a section about functional and imperative in lbmref
a3c7fda1 update lbmref with latest arithmetic performance
5eb0f0d5 added new charts to show arithmetic performance. TODO update lbmref to display correct info
4d784e27 fix bug introduced into 64bit version by recent changes

git-subtree-dir: lispBM/lispBM
git-subtree-split: 85675e29a299664697476a5b8d64bd50b450f575

+++ src/extensions/string_extensions.c
@@ -322,7 +322,7 @@ static lbm_value ext_str_replace(lbm_value *args, lbm_uint argn) {
   //    orig points to the remainder of orig after "end of rep"
   while (count--) {
     ins = strstr(orig, rep);
-    len_front = (size_t)ins - (unsigned int)orig;
+    len_front = (size_t)ins - (size_t)orig;
     tmp = strncpy(tmp, orig, len_front) + len_front;
     tmp = strcpy(tmp, with) + len_with;
     orig += len_front + len_rep; // move to next "end of rep"
